apps/prf: move Reset Results to Utilities submenu
Move the Reset Results utility (CONFIG_MFG only) from the MFG main
menu to the Utilities submenu.
