+++ src/fw/apps/prf/mfg_utilities.c
@@ -4,6 +4,7 @@
 #include "applib/app.h"
 #include "applib/ui/ui.h"
 #include "applib/ui/dialogs/confirmation_dialog.h"
+#include "apps/prf/mfg_test_result.h"
 #include "apps/prf/mfg_utilities.h"
 #include "kernel/pbl_malloc.h"
 #include "pbl/services/bluetooth/bluetooth_persistent_storage.h"
@@ -49,8 +50,46 @@ static void prv_select_clear_bondings(int index, void *context) {
   app_confirmation_dialog_push(confirmation_dialog);
 }
 
+#ifdef CONFIG_MFG
+static void prv_reset_results_confirmed(ClickRecognizerRef recognizer, void *context) {
+  ConfirmationDialog *confirmation_dialog = (ConfirmationDialog *)context;
+  confirmation_dialog_pop(confirmation_dialog);
+
+  bool confirmed = (click_recognizer_get_button_id(recognizer) == BUTTON_ID_UP);
+  if (confirmed) {
+    mfg_test_result_reset();
+  }
+}
+
+static void prv_reset_results_click_config(void *context) {
+  window_single_click_subscribe(BUTTON_ID_UP, prv_reset_results_confirmed);
+  window_single_click_subscribe(BUTTON_ID_DOWN, prv_reset_results_confirmed);
+  window_single_click_subscribe(BUTTON_ID_BACK, prv_reset_results_confirmed);
+}
+
+static void prv_select_reset_results(int index, void *context) {
+  ConfirmationDialog *confirmation_dialog = confirmation_dialog_create("Reset Results");
+  Dialog *dialog = confirmation_dialog_get_dialog(confirmation_dialog);
+
+  dialog_set_text(dialog, "Reset MFG results?\n\nThis action cannot be undone!");
+  dialog_set_background_color(dialog, GColorOrange);
+  dialog_set_text_color(dialog, GColorWhite);
+
+  confirmation_dialog_set_click_config_provider(confirmation_dialog,
+                                                prv_reset_results_click_config);
+
+  ActionBarLayer *action_bar = confirmation_dialog_get_action_bar(confirmation_dialog);
+  action_bar_layer_set_context(action_bar, confirmation_dialog);
+
+  app_confirmation_dialog_push(confirmation_dialog);
+}
+#endif
+
 static const SimpleMenuItem s_menu_items[] = {
   { .title = "Clear Bondings", .callback = prv_select_clear_bondings },
+#ifdef CONFIG_MFG
+  { .title = "Reset Results",  .callback = prv_select_reset_results },
+#endif
 };
 
 static void prv_window_load(Window *window) {
